break up resistor, eliminate uses of leddriver
