+++ edg/circuits/__init__.py
@@ -26,6 +26,8 @@ from .I2cPullup import I2cPullup
 
 from .LevelShifter import BidirectionalLevelShifter
 
+from .VoltageClamping import AnalogClampResistor, DigitalClampResistor
+
 from .OpampCircuits import OpampFollower, Amplifier, DifferentialAmplifier, IntegratorInverting, SummingAmplifier
 from .OpampCurrentSensor import OpampCurrentSensor
 from .VoltageComparator import VoltageComparator
@@ -58,3 +60,4 @@ from .ResistiveDivider import (
 
 from .RfNetworks import DiscreteRfWarning, LLowPassFilter, LLowPassFilterWith2HNotch, LHighPassFilter, PiLowPassFilter
 from .UsbBitBang import UsbBitBang
+from .UsbSeriesResistor import UsbSeriesResistor
